drm/i915/display: split i915 specific code from intel_fbdev
commit 80d20fd99124800749d605c733911a8d9da78e2b ("drm/i915/display: split i915 specific code from intel_fbdev") upstream.

Split out code from intel_fbdev that can not be share between i915 and
xe. Create new i915 specific source/header file intel_fbdev_fb.[ch] which
contains this code.

deepin-Intel-SIG: commit 80d20fd99124 ("drm/i915/display: split i915 specific code from intel_fbdev").

Link: https://patchwork.freedesktop.org/patch/msgid/20231115090719.3210079-2-jouni.hogander@intel.com
[ Junxiao Chang: amend commit log ]
